[engraving] replaced QList to std. Part 4

+++ src/engraving/layout/layoutlyrics.cpp
@@ -206,7 +206,7 @@ static void applyLyricsMin(Measure* m, int staffIdx, qreal yMin)
 void LayoutLyrics::layoutLyrics(const LayoutOptions& options, const Score* score, System* system)
 {
     std::vector<int> visibleStaves;
-    for (int staffIdx = system->firstVisibleStaff(); staffIdx < score->nstaves(); staffIdx = system->nextVisibleStaff(staffIdx)) {
+    for (size_t staffIdx = system->firstVisibleStaff(); staffIdx < score->nstaves(); staffIdx = system->nextVisibleStaff(staffIdx)) {
         visibleStaves.push_back(staffIdx);
     }
