+++ src/engraving/libmscore/barline.h
@@ -97,7 +97,7 @@ public:
 
     // Score Tree functions
     EngravingObject* scanParent() const override;
-    EngravingObject* scanChild(int idx) const override;
+    EngravingObject* scanChild(size_t idx) const override;
     size_t scanChildCount() const override;
 
     BarLine* clone() const override { return new BarLine(*this); }
