+++ src/engraving/libmscore/tuplet.cpp
@@ -816,7 +816,7 @@ Shape Tuplet::shape() const
 
 void Tuplet::scanElements(void* data, void (* func)(void*, EngravingItem*), bool all)
 {
-    for (int i = 0; i < scanChildCount(); ++i) {
+    for (size_t i = 0; i < scanChildCount(); ++i) {
         EngravingObject* child = scanChild(i);
         if (child == _number && !all) {
             continue; // don't scan number unless all is true
